WIG-87-SA * Changed from String to a value-object for /incomecar to <?> to take different values * Created a new package for value objects and klass for IncomeCar * Changed in CarServiceImpl to return a list with IncomeCar * Changed the ResponseEntity for when a choice doesn't exists to a badRequest, will change the message later * Created package for models and moved entities and valueobjects packages in there
