+++ src/main/java/com/wigell/wigellcarrental/controllers/CustomerController.java
@@ -1,19 +1,16 @@
 package com.wigell.wigellcarrental.controllers;
 
-import com.wigell.wigellcarrental.entities.Order;
-import com.wigell.wigellcarrental.entities.Car;
-import com.wigell.wigellcarrental.entities.Customer;
-import com.wigell.wigellcarrental.entities.Order;
+import com.wigell.wigellcarrental.models.entities.Order;
+import com.wigell.wigellcarrental.models.entities.Car;
+import com.wigell.wigellcarrental.models.entities.Customer;
 import com.wigell.wigellcarrental.services.CarServiceImpl;
 import com.wigell.wigellcarrental.services.CustomerServiceImpl;
-import com.wigell.wigellcarrental.services.OrderService;
 import com.wigell.wigellcarrental.services.OrderServiceImpl;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 /* import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder; */
-import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
