+++ src/main/java/com/wigell/wigellcarrental/models/entities/Car.java
@@ -1,4 +1,4 @@
-package com.wigell.wigellcarrental.entities;
+package com.wigell.wigellcarrental.models.entities;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.wigell.wigellcarrental.enums.CarStatus;
