+++ src/main/java/com/wigell/wigellcarrental/models/entities/Order.java
@@ -1,4 +1,4 @@
-package com.wigell.wigellcarrental.entities;
+package com.wigell.wigellcarrental.models.entities;
 
 import jakarta.persistence.*;
