+++ src/main/java/com/wigell/wigellcarrental/repositories/CarRepository.java
@@ -1,6 +1,6 @@
 package com.wigell.wigellcarrental.repositories;
 
-import com.wigell.wigellcarrental.entities.Car;
+import com.wigell.wigellcarrental.models.entities.Car;
 import com.wigell.wigellcarrental.enums.CarStatus;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
