+++ src/main/java/com/wigell/wigellcarrental/repositories/CustomerRepository.java
@@ -1,6 +1,6 @@
 package com.wigell.wigellcarrental.repositories;
 
-import com.wigell.wigellcarrental.entities.Customer;
+import com.wigell.wigellcarrental.models.entities.Customer;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
