+++ src/main/java/com/wigell/wigellcarrental/services/CustomerService.java
@@ -1,6 +1,6 @@
 package com.wigell.wigellcarrental.services;
 
-import com.wigell.wigellcarrental.entities.Customer;
+import com.wigell.wigellcarrental.models.entities.Customer;
 
 import java.util.List;
